+++ services/workflows-service/src/workflow/workflow.controller.external.ts
@@ -31,6 +31,7 @@ import { CurrentProject } from '@/common/decorators/current-project.decorator';
 import { EndUserService } from '@/end-user/end-user.service';
 import { WorkflowTokenService } from '@/auth/workflow-token/workflow-token.service';
 import { Public } from '@/common/decorators/public.decorator';
+import { WorkflowDefinitionService } from '@/workflow-defintion/workflow-definition.service';
 
 @swagger.ApiBearerAuth()
 @swagger.ApiTags('external/workflows')
@@ -43,6 +44,7 @@ export class WorkflowControllerExternal {
     protected readonly rolesBuilder: nestAccessControl.RolesBuilder,
     private readonly endUserService: EndUserService,
     private readonly workflowTokenService: WorkflowTokenService,
+    private readonly workflowDefinitionService: WorkflowDefinitionService,
   ) {}
 
   // GET /workflows
@@ -170,9 +172,13 @@ export class WorkflowControllerExternal {
 
     const hasSalesforceRecord =
       Boolean(body.salesforceObjectName) && Boolean(body.salesforceRecordId);
+    const latestDefinitionVersion = await this.workflowDefinitionService.getLatestVersion(
+      workflowId,
+      currentProjectId,
+    );
 
     const actionResult = await this.service.createOrUpdateWorkflowRuntime({
-      workflowDefinitionId: workflowId,
+      workflowDefinitionId: latestDefinitionVersion.id,
       context,
       config,
       projectIds,
@@ -190,7 +196,7 @@ export class WorkflowControllerExternal {
     });
   }
 
-  // POST /event
+  /// POST /event
   @common.Post('/:id/event')
   @swagger.ApiOkResponse()
   @common.HttpCode(200)