+++ services/workflows-service/src/workflow/workflow.service.ts
@@ -63,12 +63,13 @@ import {
   WorkflowRuntimeListQueryResult,
 } from './types';
 import { addPropertiesSchemaToDocument } from './utils/add-properties-schema-to-document';
-import { WorkflowDefinitionRepository } from './workflow-definition.repository';
+import { WorkflowDefinitionRepository } from '../workflow-defintion/workflow-definition.repository';
 import { WorkflowEventEmitterService } from './workflow-event-emitter.service';
 import {
   ArrayMergeOption,
   WorkflowRuntimeDataRepository,
 } from './workflow-runtime-data.repository';
+import mime from 'mime';
 
 type TEntityId = string;
 
@@ -260,7 +261,10 @@ export class WorkflowService {
         ...workflow.context,
         documents: workflow.context?.documents?.map(
           (document: DefaultContextSchema['documents'][number]) => {
-            return addPropertiesSchemaToDocument(document);
+            return addPropertiesSchemaToDocument(
+              document,
+              workflow.workflowDefinition.documentsSchema,
+            );
           },
         ),
       },
@@ -815,8 +819,8 @@ export class WorkflowService {
       id: documentId,
     };
 
-    const newDocument = addPropertiesSchemaToDocument(document);
-    const propertiesSchema = newDocument?.propertiesSchema ?? {};
+    const documentSchema = addPropertiesSchemaToDocument(document, workflowDef.documentsSchema);
+    const propertiesSchema = documentSchema?.propertiesSchema ?? {};
     if (Object.keys(propertiesSchema)?.length) {
       let propertiesSchemaForValidation = propertiesSchema;
       if (!checkRequiredFields) {
@@ -825,7 +829,7 @@ export class WorkflowService {
       }
       const validatePropertiesSchema = ajv.compile(propertiesSchemaForValidation);
 
-      const isValidPropertiesSchema = validatePropertiesSchema(newDocument?.properties);
+      const isValidPropertiesSchema = validatePropertiesSchema(documentSchema?.properties);
 
       if (!isValidPropertiesSchema) {
         throw new BadRequestException(
@@ -840,7 +844,11 @@ export class WorkflowService {
 
     const updatedWorkflow = await this.updateContextById(
       workflowId,
-      this.updateDocumentInContext(runtimeData.context, newDocument, documentsUpdateContextMethod),
+      this.updateDocumentInContext(
+        runtimeData.context,
+        documentSchema,
+        documentsUpdateContextMethod,
+      ),
       [projectId],
       documentsUpdateContextMethod === 'director' ? 'by_index' : 'by_id',
     );
@@ -884,7 +892,7 @@ export class WorkflowService {
         this.workflowEventEmitter.emit('workflow.completed', {
           runtimeData: updatedWorkflow,
           state: updatedWorkflow.state,
-          // @ts-expect-error - error from Prisma types fix
+          //@ts-expect-error
           entityId: updatedWorkflow.businessId || updatedWorkflow.endUserId,
           correlationId,
         });
@@ -1098,7 +1106,7 @@ export class WorkflowService {
       await this.workflowRuntimeDataRepository.updateById(parentMachine?.id, {
         data: {
           status: 'active',
-          // @ts-expect-error - error from Prisma types fix
+          //@ts-expect-error
           state: parentMachine?.workflowDefinition?.definition?.initial as string,
           context: {
             ...parentMachine?.context,
@@ -1668,15 +1676,33 @@ export class WorkflowService {
         const documentId = document.id! || getDocumentId(document, false);
 
         const persistedFile = await this.fileService.copyToDestinationAndCreate(
-          { id: documentId, uri: documentPage.uri, provider: documentPage.provider },
+          {
+            id: documentId,
+            uri: documentPage.uri,
+            provider: documentPage.provider,
+            // TODO: Solve once DefaultContextSchema is typed by Typebox
+            fileName: (
+              documentPage as typeof documentPage & {
+                fileName: string;
+              }
+            ).fileName,
+          },
           entityId,
           projectId,
           customerName,
         );
 
         const ballerineFileId = documentPage.ballerineFileId || persistedFile?.id;
+        const mimeType =
+          persistedFile?.mimeType ||
+          mime.getType(persistedFile.fileName || persistedFile.uri || '');
 
-        return { ...documentPage, type: persistedFile?.mimeType, ballerineFileId };
+        return {
+          ...documentPage,
+          type: mimeType,
+          ballerineFileId,
+          fileName: persistedFile?.fileName,
+        };
       }),
     );
   }
@@ -1892,7 +1918,7 @@ export class WorkflowService {
     }
 
     this.workflowEventEmitter.emit('workflow.state.changed', {
-      // @ts-expect-error - error from Prisma types fix
+      //@ts-expect-error
       entityId,
       state: updatedRuntimeData.state,
       correlationId: updatedRuntimeData.context.ballerineEntityId,