+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @ballerine/headless-example
 
+## 0.1.28
+
+### Patch Changes
+
+- Updated dependencies [340d9d4e]
+- Updated dependencies [ea423a22]
+  - @ballerine/common@0.7.29
+  - @ballerine/workflow-browser-sdk@0.5.28
+
+## 0.1.27
+
+### Patch Changes
+
+- Updated dependencies [ea423a22]
+  - @ballerine/common@0.7.28
+  - @ballerine/workflow-browser-sdk@0.5.27
+
+## 0.1.26
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies [59afd0b4]
+  - @ballerine/common@0.7.27
+  - @ballerine/workflow-browser-sdk@0.5.26
+
+## 0.1.25
+
+### Patch Changes
+
+- Updated dependencies [bfded451]
+- Updated dependencies
+  - @ballerine/common@0.7.26
+  - @ballerine/workflow-browser-sdk@0.5.25
+
 ## 0.1.24
 
 ### Patch Changes